apps/apiManager/requirements.txt
Pillow==10.2.0
portalocker==2.8.2
pystray==0.19.5
pywinauto==0.6.8
